+++ devices/sycl/sycl_conv.h
@@ -8,10 +8,10 @@
 
 OIDN_NAMESPACE_BEGIN
 
+  namespace xelp  { std::shared_ptr<Conv> newConv(const Ref<SYCLEngine>& engine, const ConvDesc& desc); }
   namespace xehpg { std::shared_ptr<Conv> newConv(const Ref<SYCLEngine>& engine, const ConvDesc& desc); }
 #if defined(__linux__)
   namespace xehpc { std::shared_ptr<Conv> newConv(const Ref<SYCLEngine>& engine, const ConvDesc& desc); }
 #endif
-  namespace gen9  { std::shared_ptr<Conv> newConv(const Ref<SYCLEngine>& engine, const ConvDesc& desc); }
 
 OIDN_NAMESPACE_END