+++ devices/sycl/sycl_conv_xelp.cpp
@@ -1,6 +1,6 @@
 // Copyright 2009-2023 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#define OIDN_ARCH_GEN9
+#define OIDN_ARCH_XELP
 
 #include "sycl_conv_gen.h"